pyproject.toml
[project]
name = "cua-test"
version = "0.1.0"
description = ""
authors = [
    {name = "James Burnside",email = "2684369+JamesBurnside@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.91.0,<2.0.0)",
    "azure-identity (>=1.23.0,<2.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
python-dotenv = "^1.1.0"

